lazy loading pages in app.tsx
